requirements.txt
#appdirs==1.4.4
asgiref==3.3.1
#CacheControl==0.12.6
#certifi==2020.12.5
#chardet==4.0.0
#colorama==0.4.4
##contextlib2==0.6.0.post1
#distlib==0.3.1
#distro==1.5.0
Django==3.1.4
django-ckeditor==6.0.0
django-js-asset==1.2.2
#html5lib==1.1
#idna==3.1
#ipaddr==2.2.0
#lockfile==0.12.2
#msgpack==1.0.2
#packaging==20.9
#pep517==0.9.1
Pillow==8.1.0
#progress==1.5
#pyparsing==2.4.7
#pytoml==0.1.21
pytz==2021.1
#requests==2.25.1
#retrying==1.3.3
six==1.15.0
sqlparse==0.4.1
#urllib3==1.26.3
#webencodings==0.5.1
